+++ src/vmm/src/rpc_interface.rs
@@ -803,6 +803,7 @@ mod tests {
     use devices::virtio::VsockError;
     use seccompiler::BpfThreadMap;
 
+    use crate::vmm_config::snapshot::{MemBackendConfig, MemBackendType};
     use mmds::data_store::MmdsVersion;
     use std::path::PathBuf;
 
@@ -1604,7 +1605,10 @@ mod tests {
         // Without resume.
         let req = VmmAction::LoadSnapshot(LoadSnapshotParams {
             snapshot_path: PathBuf::new(),
-            mem_file_path: PathBuf::new(),
+            mem_backend: MemBackendConfig {
+                backend_type: MemBackendType::File,
+                backend_path: PathBuf::new(),
+            },
             enable_diff_snapshots: false,
             resume_vm: false,
         });
@@ -1617,7 +1621,10 @@ mod tests {
         // With resume.
         let req = VmmAction::LoadSnapshot(LoadSnapshotParams {
             snapshot_path: PathBuf::new(),
-            mem_file_path: PathBuf::new(),
+            mem_backend: MemBackendConfig {
+                backend_type: MemBackendType::File,
+                backend_path: PathBuf::new(),
+            },
             enable_diff_snapshots: false,
             resume_vm: true,
         });
@@ -2025,7 +2032,10 @@ mod tests {
         check_runtime_request_err(
             VmmAction::LoadSnapshot(LoadSnapshotParams {
                 snapshot_path: PathBuf::new(),
-                mem_file_path: PathBuf::new(),
+                mem_backend: MemBackendConfig {
+                    backend_type: MemBackendType::File,
+                    backend_path: PathBuf::new(),
+                },
                 enable_diff_snapshots: false,
                 resume_vm: false,
             }),
@@ -2044,7 +2054,10 @@ mod tests {
         // Load snapshot should no longer be allowed.
         let req = VmmAction::LoadSnapshot(LoadSnapshotParams {
             snapshot_path: PathBuf::new(),
-            mem_file_path: PathBuf::new(),
+            mem_backend: MemBackendConfig {
+                backend_type: MemBackendType::File,
+                backend_path: PathBuf::new(),
+            },
             enable_diff_snapshots: false,
             resume_vm: false,
         });