+++ CHANGELOG.md
@@ -17,6 +17,18 @@
   piggy-backing on `--http-api-max-payload-size`. If left unconfigured it
   defaults to the value of `--http-api-max-payload-size`, to provide backwards
   compatibility.
+- Added optional `mem_backend` body field in `PUT` requests on `/snapshot/load`.
+  This new parameter is an object that defines the configuration of the backend
+  responsible for handling memory loading during snapshot restore. The
+  `mem_backend` parameter contains `backend_type` and `backend_path` required
+  fields. `backend_type` is an enum that can take either `File` or `Uffd` as
+  value. Interpretation of `backend_path` field depends on the value of
+  `backend_type`. If `File`, then the user must provide the path to file that
+  contains the guest memory to be loaded. Otherwise, if `backend_type` is `Uffd`,
+  then `backend_path` is the path to a unix domain socket where a custom page
+  fault handler process is listening and expecting a UFFD to be sent by
+  Firecracker. The UFFD is used to handle the guest memory page faults in the
+  separate process.
 
 ### Changed
 
@@ -25,6 +37,7 @@
 - MmdsV2 is now Generally Available.
 - MmdsV1 is now deprecated and will be removed in Firecracker v2.0.0.
   Use MmdsV2 instead.
+- Deprecated `mem_file_path` body field in `PUT` on `/snapshot/load` request.
 
 ### Fixed